user login and admin login seperated and user post list and detail list integrated

+++ src/services/admin-panel/postServices.js
@@ -1,6 +1,6 @@
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 
-import API from "./Api";
+import API from "../Api";
 import { toast } from "react-toastify";
 
 // Create a new post
@@ -9,7 +9,7 @@ export const useCreatePost = () => {
 
   return useMutation({
     mutationFn: async (data) => {
-      const response = await API.post(`post/create`, data);
+      const response = await API.post(`/admin-panel/post/create`, data);
       return response.data;
     },
     onSuccess: () => {
@@ -31,8 +31,9 @@ export const useGetAllPosts = (page = 1, limit = 10) => {
   return useQuery({
     queryKey: ["post", page, limit],
     queryFn: async () => {
-      const response = await API.get(`post/get-all`, {
+      const response = await API.get(`/admin-panel/post/get-all`, {
         params: { page, limit },
+        headers: { Authorization: `Bearer ` },
       });
       return response?.data;
     },
@@ -44,7 +45,7 @@ export const useGetPostById = (id) => {
   return useQuery({
     queryKey: ["post", id],
     queryFn: async () => {
-      const response = await API.get(`post/get-one-by-id/${id}`);
+      const response = await API.get(`/admin-panel/post/get-one-by-id/${id}`);
       return response?.data;
     },
     enabled: !!id,
@@ -57,7 +58,10 @@ export const useUpdatePost = () => {
 
   return useMutation({
     mutationFn: async ({ id, data }) => {
-      const response = await API.put(`post/update-by-id/${id}`, data);
+      const response = await API.put(
+        `/admin-panel/post/update-by-id/${id}`,
+        data
+      );
       return response.data;
     },
     onSuccess: () => {
@@ -76,7 +80,7 @@ export const useDeletePost = () => {
 
   return useMutation({
     mutationFn: async (id) => {
-      const response = await API.delete(`post/delete-by-id/${id}`);
+      const response = await API.delete(`/admin-panel/post/delete-by-id/${id}`);
       return response.data;
     },
     onSuccess: () => {
